+++ ta4j-core/src/main/java/org/ta4j/core/rules/NotRule.java
@@ -46,6 +46,7 @@ public class NotRule extends AbstractRule {
      */
     public NotRule(Rule ruleToNegate) {
         this.ruleToNegate = ruleToNegate;
+        setName(createCompositeName(getClass().getSimpleName(), ruleToNegate.getName()));
     }
 
     @Override